confirm alert added

+++ src/components/SupervisorMenu.jsx
@@ -7,6 +7,8 @@ import report from "../images/clipboard-minus.png";
 import chat from "../images/chat-typing.png";
 import logout from "../images/logout.png";
 import { Link } from "react-router-dom";
+import { confirmAlert } from "react-confirm-alert";
+import "react-confirm-alert/src/react-confirm-alert.css";
 
 function SupervisorMenu({ status }) {
 	const ActiveTab = s => {
@@ -236,9 +238,21 @@ function SupervisorMenu({ status }) {
 					src={logout}
 					alt='logout'
 					onClick={() => {
-						window.localStorage.clear();
-
-						window.location.reload(false);
+						confirmAlert({
+							message: "Are you sure to logout?",
+							buttons: [
+								{
+									label: "Yes",
+									onClick: () => {
+										window.localStorage.clear();
+										window.location.reload(false);
+									},
+								},
+								{
+									label: "No",
+								},
+							],
+						});
 					}}
 				/>
 			</div>
